requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-restql==0.11.0
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyPEG2==2.15.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
